Ok we almost setup the game logic.
The spawn should work on portrait and landscape mode.
I'm thinking on how I should setup the Fade.
Looks like the main collection should have a GUI layout to fade in out.
This should be always active.
But the game itself is almost ready.
Great the game fails on real device....
what could be then ?
We need Visual studio code.
it starts to be complex.